core/requirements.txt
Django
sorl-thumbnail
psycopg2-binary
pysolr
djangorestframework
django-oscar-api